+++ test/index.spec.ts
@@ -136,4 +136,27 @@ describe('Solidity Merger', () => {
   it('should compile file with using directive at root level (0.8.13 support)', async () => {
     await testFile('UsingDirectiveContract');
   });
+
+  it('should compile file with block.prevrandao in body (0.8.18 support)', async () => {
+    await testFile('PrevrandaoContract');
+  });
+
+  it('should compile file with user defined operators (0.8.19 support)', async () => {
+    await testFile('UserDefinedOperators');
+  });
+
+  it('should compile file with additional roots defined', async () => {
+    await testFile('ImportWithAdditionalRoot', {
+      additionalRoots: ['./test/contracts/imports'],
+    });
+  });
+
+  it('should not compile file without additional roots defined', async () => {
+    try {
+      await testFile('ImportWithAdditionalRoot', {});
+      assert.equal(true, false, 'Should never happen');
+    } catch (e) {
+      assert.isOk(e);
+    }
+  });
 });